Moved image-util.c into image.c
