--- src/lib/image/image-util.h
@@ -1,30 +1,0 @@
-/* 
- * Copyright (c) 2015-2016, Gregory M. Kurtzer. All rights reserved.
- * 
- * “Singularity” Copyright (c) 2016, The Regents of the University of California,
- * through Lawrence Berkeley National Laboratory (subject to receipt of any
- * required approvals from the U.S. Dept. of Energy).  All rights reserved.
- * 
- * This software is licensed under a customized 3-clause BSD license.  Please
- * consult LICENSE file distributed with the sources of this project regarding
- * your rights to use or distribute this software.
- * 
- * NOTICE.  This Software was developed under funding from the U.S. Department of
- * Energy and the U.S. Government consequently retains certain rights. As such,
- * the U.S. Government has been granted for itself and others acting on its
- * behalf a paid-up, nonexclusive, irrevocable, worldwide license in the Software
- * to reproduce, distribute copies to the public, prepare derivative works, and
- * perform publicly and display publicly, and to permit other to do so. 
- * 
-*/
-
-
-#ifndef __IMAGE_UTIL_H_
-#define __IMAGE_UTIL_H_
-
-    int singularity_image_check(FILE *image_fp);
-    int singularity_image_offset(FILE *image_fp);
-    int singularity_image_create(char *image, int size);
-    int singularity_image_expand(char *image, int size);
-
-#endif /* __IMAGE_UTIL_H */
